Refactored Telegram CommandBase
