+++ chat/telegram/commands/app.py
@@ -1,5 +1,5 @@
 from app.app import AyanamiApp
-from models.command_base import CommandBase
+from chat.telegram.commands.base import CommandBase
 from telegram import Update
 from telegram.ext import ContextTypes, CommandHandler
